feat: Added auth api calls in User service handlers

+++ src/main/java/com/sitblueprint/admin/service/UserServiceImpl.java
@@ -1,5 +1,6 @@
 package com.sitblueprint.admin.service;
 
+import com.sitblueprint.admin.model.AuthUser;
 import com.sitblueprint.admin.model.User;
 import com.sitblueprint.admin.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -36,11 +37,26 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User updateUser(User user) {
+        try {
+            authApiService.updateAuthUser(new AuthUser(user));
+        } catch (Exception e) {
+            throw new RuntimeException("Auth API failed to update user " + user.getUsername());
+        }
         return userRepository.saveAndFlush(user);
     }
 
     @Override
     public void deleteUserById(Long userId) {
+        Optional<User> optionalUserToDelete = userRepository.findById(userId);
+        if (optionalUserToDelete.isEmpty()) {
+            throw new RuntimeException("User with id " + userId + "was not found");
+        }
+        User userToDelete = optionalUserToDelete.get();
+        try {
+            authApiService.deleteAuthUser(userToDelete.getUsername());
+        } catch (Exception e) {
+            throw new RuntimeException("Auth API failed to delete user " + userToDelete.getId());
+        }
         userRepository.deleteById(userId);
     }
 
@@ -51,6 +67,11 @@ public class UserServiceImpl implements UserService {
             throw new RuntimeException("User not found with id " + userId);
         }
         User user = userOptional.get();
+        try {
+            authApiService.enableAuthUser(user.getUsername());
+        } catch (Exception e) {
+            throw new RuntimeException("Authentication API failed to enable user " + user.getUsername());
+        }
         user.setEnabled(true);
         userRepository.save(user);
     }
@@ -62,6 +83,11 @@ public class UserServiceImpl implements UserService {
             throw new RuntimeException("User not found with id " + userId);
         }
         User user = userOptional.get();
+        try {
+            authApiService.disableAuthUser(user.getUsername());
+        } catch (Exception e) {
+            throw new RuntimeException("Authentication API failed to disable user " + user.getUsername());
+        }
         user.setEnabled(false);
         userRepository.save(user);
     }
